pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31"
python-dotenv = "^1.0"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.0"
black = "^24.0"
isort = "^5.13"
mypy = "^1.9"

[tool.poetry.group.test.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ---------------------------------------------------------------------------
# Tool settings
# ---------------------------------------------------------------------------

# ── black ──────────────────────────────────────────────────────────────────
[tool.black]
# Длина строки
line-length = 119
# Целевая версия Python
target-version = ["py311"]
# Файлы, которые black будет форматировать
include = '\.pyi?$'
# Директории, которые black пропускает
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | venv
  | dist
  | build
  | migrations
)/
'''

# ── isort ──────────────────────────────────────────────────────────────────
[tool.isort]
# Профиль black обеспечивает полную совместимость форматов импортов
profile = "black"
line_length = 119
# Один импорт на строку
force_single_line = true
# Пакеты, относящиеся к первой стороне (собственный код)
known_first_party = ["src"]
# Пути, по которым isort ищет пакеты первой стороны
src_paths = ["src", "tests"]
# Директории, которые isort пропускает
skip = [".venv", "venv", "dist", "build", "migrations"]
# Отделять импорты стандартной библиотеки от сторонних пустой строкой
force_sort_within_sections = false
# Добавлять трейлинг-запятую в многострочных импортах (согласованно с black)
force_grid_wrap = 0
combine_as_imports = true

# ── mypy ───────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
# Строгий режим включает всё ниже разом
strict = true
# Не падать на отсутствующих стаб-пакетах сторонних библиотек
ignore_missing_imports = true
# Требовать аннотации типов для всех функций
disallow_untyped_defs = true
disallow_incomplete_defs = true
# Запрещать неявный Any
disallow_any_generics = true
# Предупреждать, если функция возвращает Any
warn_return_any = true
# Предупреждать о неиспользуемых опциях в конфиге
warn_unused_configs = true
# Предупреждать об избыточных приведениях типов
warn_redundant_casts = true
# Показывать коды ошибок в выводе — удобно для noqa-комментариев
show_error_codes = true
# Исключения из проверки
exclude = [
    "venv/",
    "migrations/",
    "build/",
    "dist/",
]

# Переопределения для тестов — чуть менее строго
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ── NOTE: flake8 ───────────────────────────────────────────────────────────
# Flake8 не поддерживает pyproject.toml нативно.
# Конфигурация находится в файле .flake8 в корне проекта.

# ── pytest ─────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra -q"
pythonpath = ["."]
cache_dir = "/tmp/.pytest_cache"

# ── coverage ───────────────────────────────────────────────────────────────
[tool.coverage.run]
branch = true
source = ["src"]
data_file = "/tmp/.coverage"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
